score-scraper/requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
click-spinner==0.1.10
colorama==0.4.6
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
drf-yasg==1.21.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
flatbuffers==23.5.26
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.13.1
ipython-genutils==0.2.0
isoduration==20.11.0
itypes==1.2.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
libclang==16.0.0
llvd==3.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
protonvpn-cli==2.2.11
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.7
pythondialog==3.5.3
pytz==2022.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyyml==0.0.2
pyzmq==25.0.2
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.3.1
scipy==1.11.1
secure-smtplib==0.1.1
selenium==4.9.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.6.2
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2022.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.20
tensorflow-intel==2.12.0